# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: Teens Party Morocco API
  description: |
    API documentation for Teens Party Morocco platform.

    ## Authentication
    Most endpoints require authentication via session cookies (NextAuth).
    Admin endpoints require elevated permissions.

    ## Rate Limiting
    - Standard users: 100 requests/minute
    - Premium users: 500 requests/minute
    - API keys: 1000 requests/minute

    ## Error Handling
    All errors follow the format:
    ```json
    {
      "success": false,
      "error": "Error message"
    }
    ```
  version: 1.0.0
  contact:
    name: Teens Party Morocco Support
    email: support@teensparty.ma
    url: https://teensparty.ma/contact
  license:
    name: Proprietary
    url: https://teensparty.ma/terms
servers:
  - url: https://teensparty.ma/api
    description: Production server
  - url: http://localhost:3000/api
    description: Development server
...
paths:
...
  /teen/profile:
    get:
      tags:
        - Teen
      summary: Get teen profile
      description: Returns the authenticated teen's profile
      security:
        - sessionAuth: []
      responses:
        '200':
          description: Teen profile
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Teen'
...
components:
  securitySchemes:
    sessionAuth:
      type: apiKey
      in: cookie
      name: next-auth.session-token
      description: NextAuth session cookie
  schemas:
...
    Teen:
      type: object
      properties:
        id:
          type: string
          format: uuid
        pseudo:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        dateOfBirth:
          type: string
          format: date
        gender:
          type: string
          enum: [male, female, other]
        avatarUrl:
          type: string
          format: uri
...